move App.jsx file to src
