+++ src/index.js
@@ -1,11 +1,12 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { BrowserRouter, Route} from 'react-router-dom';
+import React from "react";
+import ReactDOM from "react-dom";
+import { BrowserRouter, Route } from "react-router-dom";
 
-import App from './App';
+import App from "./components/App";
 
 ReactDOM.render(
-    <BrowserRouter basename={"/tf-jsx-demo/exolot/"}>
-        <Route path="/" component={App} />
-    </BrowserRouter>, document.getElementById('root'));
-
+  <BrowserRouter basename={"/tf-jsx-demo/exolot/"}>
+    <Route path="/" component={App} />
+  </BrowserRouter>,
+  document.getElementById("root")
+);
